ETL_tweet/requirements.txt
#add all the packages
pymongo
sqlalchemy
psycopg2-binary
pandas
vaderSentiment
